+++ WeenieFab/WeenieFab/ToDoList.txt
@@ -1,8 +1,11 @@
 ** To Finish Milestone 1 **
 *Priority*
 Add Generator Table
-Add Instance ID's (IID)
 Add Positions Table
+
+*Issues
+Found in Weenie 21383 Gaelan.sql
+The Last used is a double.  May just this a text field, because it really isn't use.
 
 * Polish *
 Code Cleanup/Organization
@@ -22,6 +25,7 @@ MarkDown for ES in WeenieFab
 
 
 **Completed**
+#Done-Added Instance ID's (IID)
 #Done-NOT IMPLEMENTING Destination DropDown (Create Items) (to many variables)
 #Done-Add Error Log
 #Done-Auto generate file name on save
