+++ WeenieFab/WeenieFab/TypeLists/PositionTypes.txt
@@ -1,0 +1,28 @@
+0 Undef 
+1 Location 
+2 Destination 
+3 Instantiation 
+4 Sanctuary 
+5 Home 
+6 ActivationMove 
+7 Target 
+8 LinkedPortalOne 
+9 LastPortal 
+10 PortalStorm 
+11 CrashAndTurn 
+12 PortalSummonLoc 
+13 HouseBoot 
+14 LastOutsideDeath 
+15 LinkedLifestone 
+16 LinkedPortalTwo 
+17 Save1 
+18 Save2 
+19 Save3 
+20 Save4 
+21 Save5 
+22 Save6 
+23 Save7 
+24 Save8 
+25 Save9 
+26 RelativeDestination 
+27 TeleportedCharacter
