+++ WeenieFab/WeenieFab/WeenieFab.csproj
@@ -22,6 +22,7 @@
     <None Remove="TypeLists\InstanceIDTypes.txt" />
     <None Remove="TypeLists\Int32Types.txt" />
     <None Remove="TypeLists\Int64Types.txt" />
+    <None Remove="TypeLists\PositionTypes.txt" />
     <None Remove="TypeLists\SkillTypes.txt" />
     <None Remove="TypeLists\SpellNames.txt" />
     <None Remove="TypeLists\StringTypes.txt" />
@@ -71,6 +72,9 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="TypeLists\Int64Types.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TypeLists\PositionTypes.txt">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="TypeLists\SkillTypes.txt">
